restructured for library crate
