--- src/kfn_io/lib.rs
@@ -1,6 +1,0 @@
-/// Container for the header data.
-pub mod kfn_header;
-/// Container for the kfn binary data.
-pub mod kfn_data;
-/// Container for the INI file at the end of the file.
-pub mod kfn_ini;
