+++ src/kfn_rs/lib.rs
@@ -1,0 +1,6 @@
+/// Container for the header data.
+pub mod kfn_header;
+/// Container for the kfn binary data.
+pub mod kfn_data;
+/// Container for the INI file at the end of the file.
+pub mod kfn_ini;
